Updated player laodout to disable default block and punch actions when holding items
