+++ Assets/Scenes/SampleScene.unity
@@ -5066,6 +5066,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 5799449508386820907, guid: 4e24ddcda3814db42b6532c47f7b53e1, type: 3}
+      propertyPath: m_AdditionalShaderChannelsFlag
+      value: 25
+      objectReference: {fileID: 0}
     - target: {fileID: 5799449508386820910, guid: 4e24ddcda3814db42b6532c47f7b53e1, type: 3}
       propertyPath: m_Name
       value: Game Element Canvas
