+++ Assets/Scripts/Equipment/EquipmentLoadout.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01a8872dce789fa4cbdd934a2648cac8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
